requirements.txt
asgiref==3.3.1
boto3==1.16.23
botocore==1.15.25
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.4
dj-database-url==0.5.0
Django==2.1.15
django-betterforms==1.2
django-colorfield==0.2.1
django-extra-views==0.13.0
django-heroku==0.3.1
django-storages==1.9.1
docopt==0.4.0
docutils==0.15.2
Faker==4.0.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
mailchimp==2.0.9
Pillow==7.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
raven==6.10.0
requests==2.23.0
s3transfer==0.3.3
sentry-sdk==0.14.2
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.8
whitenoise==5.0.1
